requirements.txt
PyYAML==6.0.2
pytest==8.3.5
requests==2.32.3
